feat(docs): unify the docs and wiki reader into one surface
