+++ packages/ui/src/docs/index.ts
@@ -2,3 +2,6 @@ export * from "./docs-tree";
 export * from "./doc-nav";
 export * from "./command-palette";
 export * from "./reader-persona";
+export * from "./docs-reader";
+export * from "./docs-header";
+export * from "./docs-page-actions";
