+++ packages/web/src/components/wiki/human-notes.tsx
@@ -1,14 +1,12 @@
 "use client";
 
-import { useState } from "react";
 import { toast } from "sonner";
-import { Pencil, StickyNote, X } from "lucide-react";
-import { Button } from "@repowise-dev/ui/ui/button";
+import { HumanNotes as HumanNotesShell } from "@repowise-dev/ui/wiki/human-notes";
 import { updatePageNotes } from "@/lib/api/pages";
 
 /**
- * Human-curated note pinned above a page's generated content, now editable
- * in place. Notes survive regeneration server-side.
+ * Web data wrapper around the presentational ``HumanNotes`` ui shell — owns the
+ * persistence call and toast feedback. Notes survive regeneration server-side.
  */
 export function HumanNotes({
   pageId,
@@ -17,104 +15,21 @@ export function HumanNotes({
   pageId: string;
   initialNotes: string | null;
 }) {
-  const [notes, setNotes] = useState(initialNotes);
-  const [editing, setEditing] = useState(false);
-  const [draft, setDraft] = useState(initialNotes ?? "");
-  const [saving, setSaving] = useState(false);
-
-  async function save() {
-    setSaving(true);
-    try {
-      const updated = await updatePageNotes(pageId, draft.trim() || null);
-      setNotes(updated.human_notes);
-      setDraft(updated.human_notes ?? "");
-      setEditing(false);
-      toast.success(updated.human_notes ? "Note saved" : "Note removed");
-    } catch (e) {
-      toast.error("Couldn't save note", {
-        description: e instanceof Error ? e.message : undefined,
-      });
-    } finally {
-      setSaving(false);
-    }
-  }
-
-  if (!editing && !notes) {
-    return (
-      <button
-        onClick={() => setEditing(true)}
-        className="mb-5 inline-flex items-center gap-1.5 text-xs text-[var(--color-text-tertiary)] transition-colors hover:text-[var(--color-text-secondary)]"
-      >
-        <StickyNote className="h-3.5 w-3.5" />
-        Add a note for your team — it survives regeneration
-      </button>
-    );
-  }
-
   return (
-    <div className="mb-5 rounded-lg border border-[var(--color-border-accent)] bg-[var(--color-accent-blue)]/5 px-4 py-3">
-      <div className="mb-1.5 flex items-center gap-1.5">
-        <StickyNote className="h-3.5 w-3.5 text-[var(--color-accent-blue)]" />
-        <span className="text-xs font-medium uppercase tracking-wider text-[var(--color-accent-blue)]">
-          Human Notes
-        </span>
-        {!editing && (
-          <button
-            onClick={() => setEditing(true)}
-            aria-label="Edit note"
-            className="ml-auto text-[var(--color-text-tertiary)] transition-colors hover:text-[var(--color-text-primary)]"
-          >
-            <Pencil className="h-3 w-3" />
-          </button>
-        )}
-        {editing && (
-          <button
-            onClick={() => {
-              setEditing(false);
-              setDraft(notes ?? "");
-            }}
-            aria-label="Cancel editing"
-            className="ml-auto text-[var(--color-text-tertiary)] transition-colors hover:text-[var(--color-text-primary)]"
-          >
-            <X className="h-3 w-3" />
-          </button>
-        )}
-      </div>
-      {editing ? (
-        <div className="space-y-2">
-          <textarea
-            value={draft}
-            onChange={(e) => setDraft(e.target.value)}
-            rows={4}
-            autoFocus
-            aria-label="Human note"
-            placeholder="Context the generated docs can't know — gotchas, history, plans…"
-            className="w-full resize-y rounded-md border border-[var(--color-border-default)] bg-[var(--color-bg-surface)] px-3 py-2 text-sm text-[var(--color-text-primary)] outline-none placeholder:text-[var(--color-text-tertiary)]"
-          />
-          <div className="flex items-center gap-2">
-            <Button size="sm" onClick={() => void save()} disabled={saving}>
-              {saving ? "Saving…" : "Save note"}
-            </Button>
-            {notes && (
-              <Button
-                size="sm"
-                variant="ghost"
-                disabled={saving}
-                onClick={() => {
-                  setDraft("");
-                  void save();
-                }}
-              >
-                Remove note
-              </Button>
-            )}
-          </div>
-        </div>
-      ) : (
-        <p className="whitespace-pre-wrap text-sm leading-relaxed text-[var(--color-text-secondary)]">
-          {notes}
-        </p>
-      )}
-    </div>
+    <HumanNotesShell
+      initialNotes={initialNotes}
+      onSave={async (value) => {
+        try {
+          const updated = await updatePageNotes(pageId, value);
+          toast.success(updated.human_notes ? "Note saved" : "Note removed");
+          return updated.human_notes;
+        } catch (e) {
+          toast.error("Couldn't save note", {
+            description: e instanceof Error ? e.message : undefined,
+          });
+          throw e;
+        }
+      }}
+    />
   );
 }
